# openapi.yaml
openapi: 3.0.3
info:
  title: Copilot Skillset Reviewer API
  description: |
    AI-powered code review and compliance system API

    This API provides endpoints for:
    - Loading and managing coding rules from GitHub repositories
    - Processing unified diff formats for code change analysis
    - Running comprehensive compliance checks on code changes
    - Generating actionable checklists from rule documents

    ## Authentication

    Most endpoints require GitHub App authentication via JWT tokens.

    ## Rate Limits

    API endpoints are rate-limited to prevent abuse:
    - 100 requests per minute per IP for public endpoints
    - 1000 requests per minute for authenticated GitHub App requests

    ## Error Handling

    All errors follow a consistent structured format with machine-readable error codes,
    correlation IDs for request tracing, and detailed error messages.
    
    ## Request Correlation
    
    All requests include correlation IDs (`x-correlation-id` header) for request tracing
    across distributed systems. If not provided by the client, a UUID will be generated.
  version: 0.9.0
  contact:
    name: Armoin LLC
    url: https://github.com/armoin-llc/copilot-skillset-reviewer
    email: support@armoin.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.copilot-skillset.com
    description: Production server
paths:
...
  /load-rules:
    post:
      summary: Load coding rules
      description: |
        Load coding rules from GitHub repositories or bundled guidance packs.

        Supports loading from:
        - GitHub repository `.copilot/` directories
        - Bundled guidance packs (baseline-secure, oss-apache-2, etc.)
        - Custom repository paths
      operationId: loadRules
      tags:
        - Rules
      security:
        - GitHubApp: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoadRulesRequest'
            examples:
              github_repo:
                summary: Load from GitHub repository
                value:
                  source: 'github'
                  repository: 'owner/repo'
                  path: '.copilot'
                  ref: 'main'
              bundled_pack:
                summary: Load bundled guidance pack
                value:
                  source: 'bundled'
                  pack: 'baseline-secure'
                  mode: 'merged'
      responses:
        '200':
          description: Rules loaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoadRulesResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Repository or pack not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /summarize-rules:
    post:
      summary: Generate intelligent rule summaries
      description: |
        Generate actionable checklists from rule documents with advanced parsing and categorization.

        Features:
        - Extracts rules from multiple markdown formats ([ASSERT], [REQUIRE], [CHECK], etc.)
        - Intelligent categorization by security, performance, testing, etc.
        - Priority-based sorting (critical → high → medium → low)
        - Context-aware rule processing with section header analysis
        - Duplicate detection and filtering
        - Performance caching for repeated requests
        - Statistical analysis of rule distribution

        Supported rule formats:
        - [ASSERT] directives (critical security rules)
        - [REQUIRE] directives (high-priority compliance)
        - [CHECK] directives (medium-priority testing)
        - [WARN] directives (low-priority style)
        - MUST/SHALL statements (critical compliance)
        - SHOULD/RECOMMENDED statements (high-priority compliance)
        - Categorized bullet points (security, performance, testing, etc.)

        Maximum 400 rules returned, prioritized by importance.
      operationId: summarizeRules
      tags:
        - Rules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SummarizeRulesRequest'
      responses:
        '200':
          description: Rules summarized successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SummarizeRulesResponse'
              example:
                checklist:
                  - id: "R001"
                    description: "forbid-pattern: (?i)password\\s*[:=]"
                    priority: "critical"
                    category: "security"
                    source: "Baseline Secure Coding Policy"
                  - id: "R002"
                    description: "Ensure all API endpoints have proper authentication"
                    priority: "high"
                    category: "security"
                    source: "Security Guidelines"
                  - id: "R003"
                    description: "Write unit tests for all new features"
                    priority: "high"
                    category: "testing"
                    source: "Testing Standards"
                statistics:
                  total: 42
                  byPriority:
                    critical: 5
                    high: 12
                    medium: 18
                    low: 7
                  byCategory:
                    security: 8
                    performance: 6
                    testing: 7
                    compliance: 9
                    style: 5
                    licensing: 3
                    documentation: 4
                totalItems: 42
        '400':
          description: Invalid rule documents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /normalize-diff:
    post:
      summary: Parse and normalize diff with comprehensive analysis
      description: |
        Parse unified diff format and extract detailed changed file information with advanced validation and analysis.

        Features:
        - **Multi-format Support**: Git diff, standard unified diff, automatic format detection
        - **Comprehensive Parsing**: File paths, line-by-line changes, hunk information, statistics
        - **File Operations**: Creation, deletion, rename/move detection with old/new paths
        - **Binary File Handling**: Detection and optional filtering of binary files
        - **Configurable Processing**: Strip levels, file size limits, validation strictness
        - **Line Number Tracking**: Accurate old/new line number assignment for each change
        - **Error Handling**: Detailed validation with specific error messages and recovery
        - **Performance Optimized**: Efficient parsing with configurable limits (1MB default)

        Supported diff formats:
        - Git unified diff (`diff --git a/file b/file`)
        - Standard unified diff (`--- a/file +++ b/file`)
        - Mixed formats with automatic detection

        Advanced capabilities:
        - Path normalization with configurable strip levels
        - File operation detection (new, deleted, renamed, binary)
        - Comprehensive statistics (additions, deletions, file counts)
        - Validation with strict/lenient modes
        - Edge case handling (empty files, large diffs, malformed input)
      operationId: normalizeDiff
      tags:
        - Diff Processing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NormalizeDiffRequest'
            examples:
              basic_git_diff:
                summary: Basic git diff
                value:
                  diff: |
                    diff --git a/src/file.ts b/src/file.ts
                    index 1234567..abcdefg 100644
                    --- a/src/file.ts
                    +++ b/src/file.ts
                    @@ -1,3 +1,4 @@
                     export function hello() {
                    +  console.log('debug');
                       return 'world';
                     }
              with_options:
                summary: Diff with processing options
                value:
                  diff: |
                    diff --git a/prefix/src/file.ts b/prefix/src/file.ts
                    index 1234567..abcdefg 100644
                    --- a/prefix/src/file.ts
                    +++ b/prefix/src/file.ts
                    @@ -1,1 +1,1 @@
                    -old line
                    +new line
                  strip: 1
                  maxFileSize: 2097152
                  allowBinary: false
                  strictValidation: true
      responses:
        '200':
          description: Diff normalized successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NormalizeDiffResponse'
              examples:
                successful_parse:
                  summary: Successfully parsed diff
                  value:
                    files:
                      - path: "src/file.ts"
                        additions: 1
                        deletions: 0
                        binary: false
                        renamed: false
                        created: false
                        deleted: false
                        hunks:
                          - header: "@@ -1,3 +1,4 @@"
                            oldStart: 1
                            oldLines: 3
                            newStart: 1
                            newLines: 4
                            lines:
                              - type: "context"
                                content: "export function hello() {"
                                oldLineNumber: 1
                                newLineNumber: 1
                              - type: "addition"
                                content: "  console.log('debug');"
                                newLineNumber: 2
                              - type: "context"
                                content: "  return 'world';"
                                oldLineNumber: 2
                                newLineNumber: 3
                    totalAdditions: 1
                    totalDeletions: 0
                    totalFiles: 1
                    stripLevel: 0
                    format: "git"
                    isValid: true
                    errors: []
        '400':
          description: Invalid diff format or validation failed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiError'
                  - $ref: '#/components/schemas/NormalizeDiffResponse'
              examples:
                validation_failed:
                  summary: Diff validation failed
                  value:
                    message: "Diff validation failed"
                    code: "NORMALIZE_DIFF_FAILED"
                    files: []
                    totalAdditions: 0
                    totalDeletions: 0
                    totalFiles: 0
                    stripLevel: 0
                    format: "unknown"
                    isValid: false
                    errors: ["Unsupported or unrecognized diff format"]
  /assert-compliance:
    post:
      summary: Run compliance checks
      description: |
        Run comprehensive compliance checks on code changes.

        Performs checks for:
        - Secret detection
        - License header validation  
        - Test presence verification
        - Import restrictions
        - File size limits
        - PR label validation
        - Commit message formatting
      operationId: assertCompliance
      tags:
        - Compliance
      security:
        - GitHubApp: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssertComplianceRequest'
      responses:
        '200':
          description: Compliance check completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssertComplianceResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /bundled-guidance/validate:
    get:
      summary: Validate all bundled guidance packs
      description: |
        Validates all bundled guidance packs for structural integrity and content validity.
        
        Checks for:
        - Required metadata fields (title, description)
        - File existence and accessibility
        - Non-empty content files
        - Valid pack structure
      operationId: validateBundledGuidance
      tags:
        - Bundled Guidance
      responses:
        '200':
          description: Validation completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BundledPackValidationResponse'
              example:
                valid: ["baseline-secure", "oss-apache-2", "enterprise-strict"]
                invalid:
                  - packId: "pci-dss"
                    errors: ["Pack pci-dss: file not found: bundled/pci-dss/policies/missing.md"]
                total: 6
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
components:
  securitySchemes:
    GitHubApp:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: GitHub App JWT token for repository access
...